# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: NGX Voice Sales Agent API
  description: |
    API para el agente de ventas conversacional NGX. Este sistema utiliza IA avanzada
    para automatizar y optimizar el proceso de ventas de servicios NGX, incluyendo
    NGX AGENTS ACCESS y programas de Hybrid Coaching.
    
    ## Características principales:
    - Conversaciones de voz en tiempo real con IA
    - Detección inteligente de objeciones y necesidades
    - Optimización de flujo conversacional con ML
    - Análisis predictivo de conversión
    - Sistema de caché multi-capa para <20ms de latencia
    
    ## Autenticación:
    Todos los endpoints requieren autenticación JWT Bearer token, excepto `/auth/login` y `/auth/register`.
    
  version: 1.0.0
  contact:
    name: NGX Development Team
    email: dev@ngx.com
  license:
    name: Proprietary
servers:
  - url: https://api.ngx-sales.com/v1
    description: Production server
  - url: https://staging-api.ngx-sales.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /predictive/objection/predict:
    post:
      tags:
        - predictive
      summary: Predecir objeciones
      description: Predice posibles objeciones del cliente basándose en la conversación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionRequest'
      responses:
        200:
          description: Predicción de objeciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectionPrediction'
  /predictive/needs/predict:
    post:
      tags:
        - predictive
      summary: Predecir necesidades
      description: Identifica las necesidades del cliente basándose en la conversación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionRequest'
      responses:
        200:
          description: Predicción de necesidades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeedsPrediction'
  /predictive/conversion/predict:
    post:
      tags:
        - predictive
      summary: Predecir probabilidad de conversión
      description: Calcula la probabilidad de que el cliente convierta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionRequest'
      responses:
        200:
          description: Predicción de conversión
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversionPrediction'
  /predictive/decision/optimize-flow:
    post:
      tags:
        - predictive
      summary: Optimizar flujo de conversación
      description: Genera estrategia óptima para la conversación usando el Decision Engine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizationRequest'
      responses:
        200:
          description: Estrategia optimizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationResult'

  # Qualification endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [customer, agent]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object

    # Analytics schemas
...
    PredictionRequest:
      type: object
      required:
        - conversation_id
        - messages
      properties:
        conversation_id:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        customer_profile:
          type: object
          additionalProperties: true
    ObjectionPrediction:
      type: object
      properties:
        objections:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                example: price
              probability:
                type: number
                format: float
              suggested_response:
                type: string
              confidence:
                type: number
                format: float
    NeedsPrediction:
      type: object
      properties:
        needs:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
                example: retention
              description:
                type: string
              priority:
                type: string
                enum: [high, medium, low]
              confidence:
                type: number
                format: float
    ConversionPrediction:
      type: object
      properties:
        probability:
          type: number
          format: float
          example: 0.75
        category:
          type: string
          enum: [very_low, low, medium, high, very_high]
        factors:
          type: array
          items:
            type: object
            properties:
              factor:
                type: string
              impact:
                type: string
                enum: [positive, negative, neutral]
              weight:
                type: number
                format: float
    OptimizationRequest:
      type: object
      required:
        - conversation_id
        - messages
      properties:
        conversation_id:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        current_objectives:
          type: object
          properties:
            need_satisfaction:
              type: number
              format: float
            objection_handling:
              type: number
              format: float
            conversion_progress:
              type: number
              format: float
    OptimizationResult:
      type: object
      properties:
        next_actions:
          type: array
          items:
            type: object
            properties:
              action:
                type: string
              description:
                type: string
              priority:
                type: string
                enum: [high, medium, low]
              confidence:
                type: number
                format: float
        confidence:
          type: number
          format: float
        decision_tree:
          type: object
          description: Árbol de decisión simplificado

    # Qualification schemas
...